Removed unnecessary cast to int, which might have negatively affected transfers over 2GB

+++ Tftp.Net/Transfer/TftpTransfer.cs
@@ -122,7 +122,7 @@ namespace Tftp.Net.Transfer
         {
             try
             {
-                ProposedOptions.TransferSize = (int)InputOutputStream.Length;
+                ProposedOptions.TransferSize = InputOutputStream.Length;
             }
             catch (NotSupportedException) { }
             finally
